// src/types/pokedex.ts
export interface PokedexResponse {
  results: Pokedex[];
}

export interface Pokedex {
  name: string;
  url: string;
}
// Generated by https://quicktype.io

export interface SinglePokedex {
  descriptions: Description[];
  id: number;
  is_main_series: boolean;
  name: string;
  names: Name[];
  pokemon_entries: PokemonEntry[];
  region: Region;
  version_groups: Region[];
}

export interface PokemonEntry {
  entry_number: number;
  pokemon_species: Region;
}

interface Description {
  description: string;
  language: Region;
}

interface Region {
  name: string;
  url: string;
}

interface Name {
  language: Region;
  name: string;
}